feat: remove_listing, remove_offer, inc_volume helper methods, and seperated upgrade.rs
